+++ NEWS.md
@@ -7,6 +7,7 @@
 - `pl$collect_all()` to efficiently collect a list of LazyFrames (#1598).
 - `<lazyframe>$remove()` and `<dataframe>$remove()` as a complement to
   `$filter()` (#1632).
+- New method `<expr>$is_close()` (#1637).
 
 ## polars 1.5.0